api/requirements.txt
# InsightPulse AI - Python Dependencies
# =============================================
# Install with: pip install -r requirements.txt

# Core Web Framework
streamlit>=1.32.0
fastapi>=0.111.0
uvicorn>=0.29.0

# AI / LLM
google-generativeai>=0.5.0
python-dotenv>=1.0.0

# Data / SQL Engine
duckdb>=0.10.2
pandas>=2.2.0
numpy>=1.26.4

# Machine Learning
scikit-learn>=1.4.0
scipy>=1.13.0
statsmodels>=0.14.0

# Visualization
plotly>=5.20.0

# HTTP / API
requests>=2.31.0
python-multipart>=0.0.9
pydantic>=2.7.0
aiofiles>=23.2.1
httpx>=0.27.0

# Export
kaleido>=0.2.1
fpdf2>=2.7.9
openpyxl>=3.1.2

# Text Matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0

# File parsing (PDF, Excel, etc.)
pypdf2>=3.0.0
pdfplumber>=0.10.3
